+++ src/elements/db/EntryQuery.php
@@ -919,7 +919,11 @@ class EntryQuery extends ElementQuery
         } else if (is_numeric($this->typeId)) {
             $this->typeId = [$this->typeId];
         } else if (!is_array($this->typeId) || !ArrayHelper::isNumeric($this->typeId)) {
-            throw new InvalidConfigException('Invalid typeId param value');
+            $this->typeId = (new Query())
+                ->select(['id'])
+                ->from([Table::ENTRYTYPES])
+                ->where(Db::parseParam('id', $this->typeId))
+                ->column();
         }
     }